+++ src/components/molecules/Pagination/index.stories.tsx
@@ -1,0 +1,54 @@
+/* eslint-disable react-hooks/rules-of-hooks */
+import React, { useState } from 'react';
+
+import type { Meta, StoryObj } from '@storybook/react';
+
+import { Pagination } from '.';
+
+const meta: Meta<typeof Pagination> = {
+  title: 'molecules/Pagination',
+  component: Pagination,
+  decorators: [
+    (Story) => (
+      <div
+        style={{
+          height: '50vh',
+          width: '100%',
+          padding: '20px',
+          display: 'flex',
+          flexDirection: 'column',
+          gap: '20px',
+        }}
+      >
+        <Story />
+      </div>
+    ),
+  ],
+  args: {
+    currentPage: 1,
+    itemsPerPage: 10,
+    totalItems: 503,
+    displayedCount: 9,
+  },
+};
+
+export default meta;
+
+type Story = StoryObj<typeof Pagination>;
+
+export const Default: Story = {
+  render: (args) => {
+    const [currentPage, setCurrentPage] = useState(args.currentPage);
+
+    return (
+      <>
+        <Pagination
+          {...args}
+          currentPage={currentPage}
+          onChange={setCurrentPage}
+        />
+        <Pagination displayedCount={args.displayedCount} loading />
+      </>
+    );
+  },
+};
